[OM-61944]: RB41176 SMA deal with non integer coupons

+++ com.vmturbo.market.component/src/main/java/com/vmturbo/market/cloudscaling/sma/entities/SMAInputContext.java
@@ -110,7 +110,7 @@ public class SMAInputContext {
                     oldVM.getBusinessAccountId(),
                     smaMatch.getTemplate(),
                     oldVM.getProviders(),
-                    (float)smaMatch.getDiscountedCoupons(),
+                    smaMatch.getDiscountedCoupons(),
                     oldVM.getZoneId(),
                     smaMatch.getReservedInstance() == null ?
                             SMAUtils.BOGUS_RI :
